Testing from centralized module
Each module that want to perform tests in its methods, needs to expose
a public generic object that could be imported by testing module to
bundle an object with all methods that are going to be tested.

This helps me to improve control over what is being tested or not.

Also, I don't want to modify my module exporting logic in order to have
access to internal methods that I don't want to export. So each it is solved
by each module, exporting a testing object that is composed by anything is
going to be tested.